test: extract RemoteControlApp to spec-helpers
